An unpacked node can now be filled based on how it is stored in the forest. Node storage methods for displaying nodes and writing nodes to files have been removed; instead, retreive an unpacked node and then display or write the unpacked node.
git-svn-id: svn://svn.code.sf.net/p/meddly/code/trunk@748 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/forests/mtmdd.cc
@@ -216,7 +216,7 @@ void MEDDLY::mtmdd_forest::sifting(int var, int top, int bottom)
 
   MEDDLY_DCASSERT(level <= top && level >= bottom);
 
-  int num = getCurrentNumNodes();
+  long num = getCurrentNumNodes();
   if(level <= (top + bottom) / 2) {
     // Move to the bottom
     while(level > bottom) {
@@ -304,7 +304,7 @@ void MEDDLY::mtmdd_forest::sifting(int var, int top, int bottom)
 
   MEDDLY_DCASSERT(getCurrentNumNodes() <= num);
   if(getCurrentNumNodes() > num) {
-    printf("Error: %d > %d\n", getCurrentNumNodes(), num);
+    printf("Error: %ld > %ld\n", getCurrentNumNodes(), num);
   }
 }
